code improving

+++ script.sh
@@ -1,13 +1,9 @@
 #!/bin/bash
 
-function mountdisks()
-{
+function mountdisks() {
     # 定义磁盘设备列表和对应的挂载点列表
-    mkdir /data
-    mkdir /data1
-    mkdir /ssd
-    devices=("/dev/sda" "/dev/sdb" "/dev/nvme1n1")  # 替换为你的磁盘设备列表
-    mount_points=("/data" "/data1" "/ssd")  # 替换为你的挂载点列表
+    declare -a devices=("/dev/sda" "/dev/sdb" "/dev/nvme1n1")  # 替换为你的磁盘设备列表
+    declare -a mount_points=("/data" "/data1" "/ssd")  # 替换为你的挂载点列表
 
     # 确保设备列表和挂载点列表长度相同
     if [ ${#devices[@]} -ne ${#mount_points[@]} ]; then
@@ -17,62 +13,80 @@ function mountdisks()
 
     # 循环处理每个磁盘
     for ((i=0; i<${#devices[@]}; i++)); do
+        # 格式化硬盘为ext4格式，适用于初次执行
         # mkfs -t ext4 ${devices[$i]}
-        device_to_mount="${devices[$i]}"
+
+        device="${devices[$i]}"
         mount_point="${mount_points[$i]}"
-        mount ${devices[$i]} ${mount_points[$i]}
-        # 获取磁盘UUID
-        uuid=$(blkid -s UUID -o value "$device_to_mount")
 
-        if [ -n "$uuid" ]; then
-            # 备份/etc/fstab文件
-            cp /etc/fstab /etc/fstab.bak
+        # 确保挂载点存在
+        mkdir -p "$mount_point"
 
-            # 添加挂载项到/etc/fstab
-            echo "UUID=$uuid $mount_point ext4 defaults 0 0" >> /etc/fstab
-            echo "Added entry to /etc/fstab for $device_to_mount with UUID $uuid"
+        # 检查磁盘是否已经挂载
+        if grep -qs "$device" /proc/mounts; then
+            echo "$device is already mounted."
         else
-            echo "Failed to get UUID for $device_to_mount"
+            # 检查是否为 ext4 文件系统，如果不是，可以根据需要添加其他文件系统类型
+            if blkid -s TYPE -o value "$device" | grep -q "^ext4$"; then
+                # 获取磁盘UUID
+                uuid=$(blkid -s UUID -o value "$device")
+
+                if [ -n "$uuid" ]; then
+                    # 备份/etc/fstab文件
+                    cp /etc/fstab /etc/fstab.bak
+
+                    # 添加挂载项到/etc/fstab
+                    echo "UUID=$uuid $mount_point ext4 defaults 0 0" >> /etc/fstab
+                    mount "$device" "$mount_point"
+                    echo "Mounted $device at $mount_point with UUID $uuid"
+                else
+                    echo "Failed to get UUID for $device"
+                fi
+            else
+                echo "$device is not formatted as ext4. Please format it accordingly."
+            fi
         fi
     done
 }
 
-function installdocker()
-{
+function installdocker() {
     # Add Docker's official GPG key:
     sudo apt-get update
-    sudo apt-get install ca-certificates curl gnupg
+    sudo apt-get install -y ca-certificates curl gnupg
+
+    local docker_keyring="/etc/apt/keyrings/docker.gpg"
+    local docker_repository="https://download.docker.com/linux/ubuntu"
+    local docker_architecture="$(dpkg --print-architecture)"
+    local docker_codename
+    docker_codename=$(. /etc/os-release && echo "$VERSION_CODENAME")
+
     sudo install -m 0755 -d /etc/apt/keyrings
-    curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo gpg --dearmor -o /etc/apt/keyrings/docker.gpg
-    sudo chmod a+r /etc/apt/keyrings/docker.gpg
+    curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo gpg --dearmor -o "$docker_keyring"
+    sudo chmod a+r "$docker_keyring"
 
     # Add the repository to Apt sources:
-    echo \
-    "deb [arch="$(dpkg --print-architecture)" signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/ubuntu \
-    "$(. /etc/os-release && echo "$VERSION_CODENAME")" stable" | \
-    sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
-    sudo apt-get update
+    echo "deb [arch=${docker_architecture} signed-by=${docker_keyring}] ${docker_repository} ${docker_codename} stable" | \
+        sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
 
+    sudo apt-get update
     sudo apt-get install -y docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin
 
     echo "Docker installation and configuration completed."
 }
 
-function configdocker()
-{
+function configdocker() {
     # Define the file path and content
-    file_path="/etc/docker/daemon.json"
-    content='{
+    local file_path="/etc/docker/daemon.json"
+    local content='{
         "data-root": "/ssd/docker",
         "exec-opts": ["native.cgroupdriver=cgroupfs"]
     }'
+
     # Check if the file already exists, and create it if not
-    if [ ! -f "$file_path" ]; then
-        sudo touch "$file_path"
-    fi
+    [ ! -f "$file_path" ] && sudo touch "$file_path"
 
     # Write the content to the file
-    echo "$content" | sudo tee "$file_path"
+    echo "$content" | sudo tee "$file_path" > /dev/null
 
     echo "Configuration file '$file_path' has been created and updated."
 
@@ -80,57 +94,66 @@ function configdocker()
     sudo systemctl daemon-reload
     sudo systemctl restart docker
 
-    distribution=$(. /etc/os-release;echo $ID$VERSION_ID)
-    curl -s -L https://nvidia.github.io/nvidia-docker/gpgkey | sudo apt-key add -
-    curl -s -L https://nvidia.github.io/nvidia-docker/$distribution/nvidia-docker.list | sudo tee /etc/apt/sources.list.d/nvidia-docker.list
+    # Add NVIDIA Docker support
+    local distribution=$(. /etc/os-release && echo "$ID$VERSION_ID")
+    local nvidia_key_url="https://nvidia.github.io/nvidia-docker/gpgkey"
+    local nvidia_repo_url="https://nvidia.github.io/nvidia-docker/${distribution}/nvidia-docker.list"
+
+    curl -s -L "$nvidia_key_url" | sudo apt-key add -
+    curl -s -L "$nvidia_repo_url" | sudo tee /etc/apt/sources.list.d/nvidia-docker.list > /dev/null
+
     sudo apt-get update && sudo apt-get install -y nvidia-container-toolkit
+
+    # Backup NVIDIA Docker repository configuration
     mv /etc/apt/sources.list.d/nvidia-docker.list nvidia-docker.list.bak
+
+    # Restart Docker for changes to take effect
     sudo systemctl restart docker
+
+    echo "NVIDIA Docker support has been configured."
 }
 
-function configssh()
-{
+function configssh() {
     # 配置文件路径
     config_file="/etc/ssh/sshd_config"
 
-    # 检查是否存在MaxSessions行
-    if grep -q "^MaxSessions" "$config_file"; then
-        # 如果存在MaxSessions行，使用sed命令修改行，去掉注释符号并追加配置
-        sed -i '/^MaxSessions/ s/#\?MaxSessions.*/MaxSessions 1000/' "$config_file"
-    else
-        # 如果不存在MaxSessions行，直接追加配置
-        echo "MaxSessions 1000" >> "$config_file"
-    fi
+    # 定义要设置的参数和值
+    max_sessions="MaxSessions"
+    max_startups="MaxStartups"
+    new_value="1000"
 
-    # 检查是否存在MaxStartups行
-    if grep -q "^MaxStartups" "$config_file"; then
-        # 如果存在MaxStartups行，使用sed命令修改行，去掉注释符号并追加配置
-        sed -i '/^MaxStartups/ s/#\?MaxStartups.*/MaxStartups 1000/' "$config_file"
-    else
-        # 如果不存在MaxStartups行，直接追加配置
-        echo "MaxStartups 1000" >> "$config_file"
-    fi
+    # 函数用于检查并更新配置行
+    update_config() {
+        local parameter="$1"
+        if grep -q "^$parameter" "$config_file"; then
+            # 如果存在该参数，使用sed命令修改行，去掉注释符号并追加配置
+            sed -i "/^$parameter/ s/#\?$parameter.*/$parameter $new_value/" "$config_file"
+        else
+            # 如果不存在该参数，直接追加配置
+            echo "$parameter $new_value" >> "$config_file"
+        fi
+    }
+
+    # 检查并更新MaxSessions配置
+    update_config "$max_sessions"
+
+    # 检查并更新MaxStartups配置
+    update_config "$max_startups"
 
     # 重启SSH服务
-    systemctl restart ssh
+    sudo systemctl restart ssh
 
-    echo "MaxSessions has been set $new_max_sessions"
-    echo "MaxStartups has been set $new_max_startups"
+    echo "SSH configuration has been updated:"
+    echo "$max_sessions set to $new_value"
+    echo "$max_startups set to $new_value"
 }
 
-function configdock()
-{
-    # 检查是否有足够权限执行
-    if [ "$EUID" -ne 0 ]; then
-        echo "请使用sudo运行此脚本"
-        exit
-    fi
-
+function configdock() {
     # 设置可执行文件路径
     script_path="/usr/local/bin/dock"
 
     # 创建可执行文件
-    cat << EOF > "$script_path"
+    cat << 'EOF' > "$script_path"
 #!/bin/bash
 
 RED='\033[0;31m'
@@ -140,10 +163,14 @@ RESET='\033[0m'
 
 read -p "请输入PID: " pid
 
-container_id=\$(cat /proc/\$pid/cgroup | grep -oP '/docker/\K.{12}')
-container_name=\$(docker ps --no-trunc --format '{{.Names}}' --filter "id=\$container_id")
+container_id=$(cat /proc/$pid/cgroup | grep -oP '/docker/\K.{12}')
+container_name=$(docker ps --no-trunc --format '{{.Names}}' --filter "id=$container_id")
 
-echo -e "PID \${RED}\$pid\${RESET} 所属的容器名称为: \${GREEN}\$container_name\${RESET}"
+if [ -n "$container_name" ]; then
+    echo -e "PID ${RED}$pid${RESET} 所属的容器名称为: ${GREEN}$container_name${RESET}"
+else
+    echo -e "PID ${RED}$pid${RESET} 不属于任何容器."
+fi
 EOF
 
     # 添加执行权限
@@ -152,21 +179,31 @@ EOF
     echo "Script has been installed and created at $script_path, now you can run command 'dock'."
 }
 
-function installapps()
-{
-    echo "Installing vim..."
-    apt install -y vim
-    echo "vim installed."
 
-    echo "Installing htop..."
-    apt install -y htop
-    echo "thop installed."
+function install_apps() {
+    # Function to check if a package is installed
+    is_package_installed() {
+        dpkg -l "$1" | grep -qE "^ii\s+$1"
+    }
 
-    echo "Installing git..."
-    apt install -y git
-    echo "git installed."
+    # Install a package if it is not already installed
+    install_package() {
+        local package_name="$1"
+        if ! is_package_installed "$package_name"; then
+            echo "Installing $package_name..."
+            sudo apt-get update
+            sudo apt-get install -y "$package_name"
+        else
+            echo "$package_name already installed."
+        fi
+    }
 
-    # 检查是否已安装 pip3
+    # Install vim, htop, git
+    install_package vim
+    install_package htop
+    install_package git
+
+    # Install pip3 if not installed
     if ! command -v pip3 &> /dev/null; then
         echo "pip3 not installed, installing..."
         sudo apt update
@@ -174,58 +211,43 @@ function installapps()
     else
         echo "pip3 installed."
     fi
+
+    # Upgrade nvitop
     pip3 install --upgrade nvitop
     echo "nvitop installed."
 
-    # Function to check if a package is installed
-    is_package_installed() {
-        dpkg -l "$1" | grep -qE "^ii\s+$1"
-    }
+    # Install curl and wget
+    install_package curl
+    install_package wget
 
-    # Check if curl is installed
-    if ! is_package_installed curl; then
-        echo "Installing curl..."
-        sudo apt-get update
-        sudo apt-get install -y curl
-    fi
-
-    # Check if wget is installed
-    if ! is_package_installed wget; then
-        echo "Installing wget..."
-        sudo apt-get update
-        sudo apt-get install -y wget
-    fi
+    # Install V2RayA dependencies
+    install_package apt-transport-https
 
     # Install V2RayA
-    curl -Ls https://mirrors.v2raya.org/go.sh | sudo bash
-    sudo systemctl disable v2ray --now
+    if ! is_package_installed v2raya; then
+        echo "Installing V2RayA..."
+        curl -Ls https://mirrors.v2raya.org/go.sh | sudo bash
 
-    # Check if wget is installed
-    if ! is_package_installed apt-transport-https; then
-        echo "Installing apt-transport-https..."
-        sudo apt-get update
-        sudo apt-get install -y apt-transport-https
+        # Add the V2RayA repository
+        wget -qO - https://apt.v2raya.org/key/public-key.asc | sudo tee /etc/apt/keyrings/v2raya.asc
+        echo "deb [signed-by=/etc/apt/keyrings/v2raya.asc] https://apt.v2raya.org/ v2raya main" | sudo tee /etc/apt/sources.list.d/v2raya.list
+
+        # Update package lists
+        sudo apt update
+
+        # Install V2RayA
+        sudo apt install -y v2raya
+        echo "V2RayA installed."
+    else
+        echo "V2RayA already installed."
     fi
-
-    # Add the V2RayA repository
-    wget -qO - https://apt.v2raya.org/key/public-key.asc | sudo tee /etc/apt/keyrings/v2raya.asc
-    echo "deb [signed-by=/etc/apt/keyrings/v2raya.asc] https://apt.v2raya.org/ v2raya main" | sudo tee /etc/apt/sources.list.d/v2raya.list
-
-    # Update package lists
-    sudo apt update
-
-    # Install V2RayA
-    sudo apt install v2raya
-
-    echo "v2raya installed."
 }
 
-function createusers()
-{
+function create_users() {
     # Specify the list of usernames
-    usernames=("hub" "chendh" "chenq" "gaowc" "lijw" "lixx" "maopc" "tangps" "wangd" "xuy" "yangcx" "zhangt")
+    local usernames=("hub" "huxg" "chendh" "chenq" "gaowc" "lijw" "lixx" "maopc" "tangps" "wangd" "wangxy" "xuy" "yangcx" "zhangt" "zouyt")
     # Common password for all users
-    password="1341454644"
+    local password="1341454644"
 
     for username in "${usernames[@]}"; do
         # Check if the user already exists
@@ -239,47 +261,50 @@ function createusers()
     done
 }
 
-function configusers()
-{
+function config_users() {
     # Specify the list of usernames
-    usernames=("hub" "huxg" "chendh" "chenq" "gaowc" "lijw" "lixx" "maopc" "tangps" "wangd" "wangxy" "xuy" "yangcx" "zhangt" "zouyt")
+    local usernames=("hub" "huxg" "chendh" "chenq" "gaowc" "lijw" "lixx" "maopc" "tangps" "wangd" "wangxy" "xuy" "yangcx" "zhangt" "zouyt" "user")
 
     for username in "${usernames[@]}"; do
-        usermod -aG docker $username
-        echo "Added $username to Docker group"
+        if id "$username" &>/dev/null; then
+            usermod -aG docker "$username"
+            echo "Added $username to Docker group"
+        else
+            echo "User $username not found."
+        fi
     done
 }
 
 function main()
 {
-    # echo "Mounting disks..."
+    echo "Mounting disks..."
     # mountdisks
-    # echo "Completed"
+    echo "Completed"
 
-    # echo "Installing Docker engine..."
-    # installdocker
-    # echo "Completed"
+    echo "Installing Docker engine..."
+    installdocker
+    echo "Completed"
 
-    # echo "Editing Docker configuration file..."
-    # configdocker
-    # echo "Completed"
+    echo "Editing Docker configuration file..."
+    configdocker
+    echo "Completed"
 
-    # echo "Editing ssh configuration file..."
-    # configssh
-    # echo "Completed"
+    echo "Editing ssh configuration file..."
+    configssh
+    echo "Completed"
 
     echo "Creating dock command..."
     configdock
     echo "Completed"
 
-    # echo "Installing some apps..."
-    # installapps
-    # echo "Completed"
+    echo "Installing some apps..."
+    installapps
+    echo "Completed"
 
-    # echo "Creating and Editing users..."
-    # createusers
-    # configusers
-    # echo "Completed"
+    echo "Creating and Editing users..."
+    createusers
+    configusers
+    echo "Completed"
 }
 
 main
